initial commit
So far the works on the semi-finished robot.
it drives (feeder forward)
it shoots (motor speed set manually from smart dashboard)
it feeds (you can toggle feed mode on and off. when in feed mode, it
will stop the wheels once you're posessing and start them once you're
not)
it passes (hold the pass button and it will pass)